fix(apigatewayv2-authorizers): update integration test for HttpLambdaAuthorizer
