+++ packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/asset.0793b3c94d0be9d628975dbbcd1dc523196de50a41fb1920880248ae3c24259b.handler/index.js
@@ -1,0 +1,13 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.handler = void 0;
+const handler = async () => {
+    return {
+        statusCode: 200,
+        body: JSON.stringify({ message: 'Hello from authenticated lambda' }),
+        headers: {
+            'Content-Type': 'application/json',
+        },
+    };
+};
+exports.handler = handler;
