--- packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/asset.54deaef2af5b9afbfc9cbcbb9261b1c0d4cce6560831d7ae1959f3da899011c8.handler/index.d.ts
@@ -1,7 +1,0 @@
-export declare const handler: () => Promise<{
-    statusCode: number;
-    body: string;
-    headers: {
-        'Content-Type': string;
-    };
-}>;
