--- packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/asset.54deaef2af5b9afbfc9cbcbb9261b1c0d4cce6560831d7ae1959f3da899011c8.handler/index.ts
@@ -1,9 +1,0 @@
-export const handler = async () => {
-  return {
-    statusCode: 200,
-    body: JSON.stringify({ message: 'Hello from authenticated lambda' }),
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  };
-};
