+++ packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/asset.9b3143f6d2fe3103f2f3162c221163ad6a5a0ba2219cc1e5af2e300325588714/index.js
@@ -1,0 +1,11 @@
+"use strict";
+/* eslint-disable no-console */
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.handler = void 0;
+const handler = async (event) => {
+    const key = event.headers['x-api-key'];
+    return {
+        isAuthorized: key === '123',
+    };
+};
+exports.handler = handler;
