--- packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/asset.d7d3785243d748927f2a8d6edcecf909f96191df27a815e305aaeba98bcd2c64/index.d.ts
@@ -1,3 +1,0 @@
-export declare const handler: (event: import("aws-lambda").APIGatewayProxyEventV2) => Promise<{
-    isAuthorized: boolean;
-}>;
