--- packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/asset.d7d3785243d748927f2a8d6edcecf909f96191df27a815e305aaeba98bcd2c64/index.ts
@@ -1,9 +1,0 @@
-/* eslint-disable no-console */
-
-export const handler = async (event: AWSLambda.APIGatewayProxyEventV2) => {
-  const key = event.headers['x-api-key'];
-
-  return {
-    isAuthorized: key === '123',
-  };
-};
