+++ packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.js.snapshot/cdk.out
@@ -1,1 +1,1 @@
-{"version":"34.0.0"}
+{"version":"36.0.0"}
