+++ packages/@aws-cdk-testing/framework-integ/test/aws-apigatewayv2-authorizers/test/http/integ.lambda.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { HttpApi, HttpMethod, HttpRoute, HttpRouteKey } from 'aws-cdk-lib/aws-apigatewayv2';
+import { AuthorizerPayloadVersion, HttpApi, HttpMethod, HttpRoute, HttpRouteKey } from 'aws-cdk-lib/aws-apigatewayv2';
 import { HttpLambdaIntegration } from 'aws-cdk-lib/aws-apigatewayv2-integrations';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import { App, Stack, CfnOutput } from 'aws-cdk-lib';
@@ -21,19 +21,12 @@ const authHandler = new lambda.Function(stack, 'auth-function', {
   code: lambda.Code.fromAsset(path.join(__dirname, '..', 'auth-handler'), { exclude: ['*.ts'] }),
 });
 
-const authorizer = new HttpLambdaAuthorizer('LambdaAuthorizer', authHandler, {
-  authorizerName: 'my-simple-authorizer',
-  identitySource: ['$request.header.X-API-Key'],
-  responseTypes: [HttpLambdaResponseType.SIMPLE],
-});
-
 const defaultAuthorizer = new HttpLambdaAuthorizer('LambdaDefaultAuthorizer', authHandler, {
   authorizerName: 'my-simple-authorizer',
   identitySource: ['$request.header.X-API-Key'],
-  responseTypes: [HttpLambdaResponseType.SIMPLE],
+  responseType: HttpLambdaResponseType.SIMPLE,
 });
 
-const httpApi = new HttpApi(stack, 'MyHttpApi');
 const httpApiWithDefaultAuthorizer = new HttpApi(stack, 'MyHttpApiWithDefaultAuthorizer', {
   defaultAuthorizer,
 });
@@ -44,22 +37,45 @@ const handler = new lambda.Function(stack, 'lambda', {
   code: lambda.AssetCode.fromAsset(path.join(__dirname, '..', 'integ.lambda.handler'), { exclude: ['*.ts'] }),
 });
 
-httpApi.addRoutes({
-  path: '/',
-  methods: [HttpMethod.GET],
-  integration: new HttpLambdaIntegration('RootIntegration', handler),
-  authorizer,
-});
-
 new HttpRoute(stack, 'Route', {
   httpApi: httpApiWithDefaultAuthorizer,
   routeKey: HttpRouteKey.with('/v1/mything/{proxy+}', HttpMethod.ANY),
   integration: new HttpLambdaIntegration('RootIntegration', handler),
 });
 
-new CfnOutput(stack, 'URL', {
-  value: httpApi.url!,
-});
 new CfnOutput(stack, 'URLWithDefaultAuthorizer', {
   value: httpApiWithDefaultAuthorizer.url!,
+});
+
+const attributes: [AuthorizerPayloadVersion | undefined, HttpLambdaResponseType | undefined][] = [
+  [undefined, undefined],
+  [undefined, HttpLambdaResponseType.SIMPLE],
+  [undefined, HttpLambdaResponseType.IAM],
+  [AuthorizerPayloadVersion.VERSION_1_0, undefined],
+  [AuthorizerPayloadVersion.VERSION_1_0, HttpLambdaResponseType.IAM],
+  [AuthorizerPayloadVersion.VERSION_2_0, undefined],
+  [AuthorizerPayloadVersion.VERSION_2_0, HttpLambdaResponseType.SIMPLE],
+  [AuthorizerPayloadVersion.VERSION_2_0, HttpLambdaResponseType.IAM],
+];
+
+attributes.forEach(([payloadFormatVersion, responseType], index) => {
+  const authorizer = new HttpLambdaAuthorizer(`LambdaAuthorizer${index}`, authHandler, {
+    authorizerName: `my-simple-authorizer${index}`,
+    identitySource: ['$request.header.X-API-Key'],
+    payloadFormatVersion,
+    responseType,
+  });
+
+  const httpApi = new HttpApi(stack, `MyHttpApi${index}`);
+
+  httpApi.addRoutes({
+    path: '/',
+    methods: [HttpMethod.GET],
+    integration: new HttpLambdaIntegration(`RootIntegration${index}`, handler),
+    authorizer,
+  });
+
+  new CfnOutput(stack, `URL${index}`, {
+    value: httpApi.url!,
+  });
 });
